requirements-lock.txt
altair==6.0.0
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
attrs==26.1.0
black==26.3.1
blinker==1.9.0
branca==0.8.2
cachetools==7.0.5
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
comm==0.2.3
contourpy==1.3.3
cryptography==47.0.0
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.18.0
executing==2.2.1
flatbuffers==25.9.23
folium==0.20.0
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jh2==5.0.11
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mypy_extensions==1.1.0
narwhals==2.19.0
nest-asyncio==1.6.0
niquests==3.18.6
numpy==2.4.4
openmeteo_requests==1.7.5
openmeteo_sdk==1.26.0
packaging==26.1
pandas==3.0.2
parso==0.8.6
pathspec==1.0.4
patsy==1.0.2
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.6
plotly==6.7.0
pluggy==1.6.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
pydeck==0.9.1
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
pytokens==0.4.1
pyzmq==27.1.0
qh3==1.7.1
referencing==0.37.0
requests==2.33.1
requests-cache==1.3.1
retry-requests==2.0.0
rpds-py==0.30.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.6
streamlit==1.56.0
streamlit-folium==0.27.1
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.6.3
urllib3-future==2.19.908
wassima==2.0.6
wcwidth==0.6.0
websockets==16.0
xyzservices==2026.3.0
